requirement.txt
streamlit
pandas
newsapi-python
scikit-learn
matplotlib
seaborn
python-docx